passes run process on route
